Proofreading done
Please double-check that where I've deleted 'tooling' is correct, I think it's a repeated word but I want to make sure I'm not interfering with technical terms. Also, for 'You can install the CLI tool by cloning running go install', should it be either 'cloning' or 'running', or is it appropriate to use both? Again, I don't want to interfere with technical terms. If you just need one or the other then please delete appropriately or let me know and I'll be happy to do it. :)