Proofreading done
All done on this one! Just a couple of questions again to make sure I'm not changing the meaning of anything technical. As with the  blurb, feel free to either make the changes yourself or ask me and I'll happily do them. 

I just want to highlight that I changed 'cypher-text' to 'cipher text' as it was written that way everywhere else. Is that okay or is 'cypher-text' a different thing? Likewise, I noticed an instance of 'cipher_text', which I've left as is just in case. 

'Message storage returns a hash of the contents of the payload which is used as the URL address.' I think this needs a comma for clarity, but placement depends on the meaning. If you mean message storage returns a hash of the payload contents, and then that hash is used as the URL address, you need a comma between 'payload' and 'which'. Currently, without a comma, it could mean the payload is used as the URL address. (I have no idea which of those is a real concept or if one's impossible, sorry!)

'A chat is generated after a handshake is complete in storage it contains a randomly generated id for the chat.' I would recommend breaking this up into two sentences, but again it depends on the meaning. Is the chat generated after the handshake is complete in storage, or does storage contain a randomly generated id for the chat? 

It'll be ''A chat is generated after a handshake is complete in storage. It contains a randomly generated id for the chat.' or 'A chat is generated after a handshake is complete. In storage, it contains a randomly generated id for the chat.' respectively. Or even 'A chat is generated. After a handshake is complete in storage, it contains a randomly generated id for the chat.'

Apologies for all the queries but I hope they make sense.